requirements.txt
django>=5.2
python-decouple
Pillow==11.3.0
psycopg2-binary
django-allauth==0.51.0
django-guardian==2.4.0
django-two-factor-auth==1.15.0
django-imagekit==4.1.0
django-storages==1.13.2
boto3==1.26.137
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
drf-spectacular==0.26.2
celery==5.2.7
redis==4.5.5
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
Pillow
django-extensions
# Add django-redis if you plan to use Redis as cache backend
django-redis
# For API filtering
django-filter
# Optional: reCAPTCHA support
django-recaptcha
# Alternatively, some distributions use `django_recaptcha` as the package name
# django_recaptcha